Add enemy detection and collision
That was a bit more fun than expected. The tutorial didn't mention that
you need to add the `move_and_slide` call to the physics loop for
object collision with trees etc, but that makes sense to have that.

Added a nice helper to work out the direction the enemy is moving
towards so that it plays the right animation as well