Add enemy attack
deviate from the tutorial here and use signals instead of the backwards
handling of attacks from the tutorial. In the original the player object
would track if the enemy is attacking. Instead here we setup a signal
from the enemy to send attack, the player handles this
The attack is on a timer which stops when the player leaves the range of
the enemy